fix(tw): the ontology tree is value/Ontology.vue, drawn only for a value that carries its ancestors
