+++ apps/catalogue/app/pages/[catalogue]/[resourceType]/[resource]/collection-events/[collectionevent].vue
@@ -16,7 +16,7 @@ import BreadCrumbs from "../../../../../../../tailwind-components/app/components
 import SideNavigation from "../../../../../components/SideNavigation.vue";
 import ContentBlocks from "../../../../../../../tailwind-components/app/components/content/ContentBlocks.vue";
 import ContentBlock from "../../../../../../../tailwind-components/app/components/content/ContentBlock.vue";
-import DisplayOntology from "../../../../../../../tailwind-components/app/components/display/Ontology.vue";
+import ValueOntology from "../../../../../../../tailwind-components/app/components/value/Ontology.vue";
 import CatalogueItemList from "../../../../../components/CatalogueItemList.vue";
 import type { Crumb } from "../../../../../../../tailwind-components/types/types";
 
@@ -172,7 +172,7 @@ useHead({
           id="sample_categories"
           title="Sample categories"
         >
-          <DisplayOntology
+          <ValueOntology
             :value="collectionEvent.sampleCategories"
             :collapse-all="false"
           />
@@ -211,7 +211,7 @@ useHead({
           id="data_categories"
           title="Data categories"
         >
-          <DisplayOntology
+          <ValueOntology
             :value="collectionEvent.dataCategories"
             :collapse-all="false"
           />
@@ -221,7 +221,7 @@ useHead({
           id="areas_of_information"
           title="Areas of information"
         >
-          <DisplayOntology
+          <ValueOntology
             :value="collectionEvent.areasOfInformation"
             :collapse-all="false"
           />
@@ -231,7 +231,7 @@ useHead({
           id="standardized_tools"
           title="Standardized tools"
         >
-          <DisplayOntology
+          <ValueOntology
             :value="collectionEvent.standardizedTools"
             :collapse-all="false"
           />
