+++ apps/tailwind-components/app/components/value/OntologyNode.vue
