+++ apps/ui/app/pages/[schema]/[table]/[entity].vue
@@ -20,6 +20,7 @@ import PageHeader from "../../../../../tailwind-components/app/components/PageHe
 import CellDetailModal from "../../../../../tailwind-components/app/components/table/cellDetail/CellDetailModal.vue";
 import ValueEMX2 from "../../../../../tailwind-components/app/components/value/EMX2.vue";
 import fetchRowData from "../../../../../tailwind-components/app/composables/fetchRowData";
+import { withOntologyAncestry } from "../../../../../tailwind-components/app/composables/fetchOntologyAncestry";
 import fetchTableMetadata from "../../../../../tailwind-components/app/composables/fetchTableMetadata";
 import { useSession } from "../../../../../tailwind-components/app/composables/useSession";
 import { useTablePermission } from "../../../../../tailwind-components/app/composables/useTablePermission";
@@ -51,7 +52,10 @@ const { isAdmin, session } = await useSession(schemaId);
 const tableMetadata = await fetchTableMetadata(schemaId, tableId);
 const { data: rowData, refresh } = await useAsyncData(
   keys || JSON.stringify(entityKeysObject),
-  () => fetchRowData(schemaId, tableId, entityKeysObject)
+  () =>
+    fetchRowData(schemaId, tableId, entityKeysObject).then((row) =>
+      withOntologyAncestry(row, tableMetadata.columns)
+    )
 );
 
 const sections = computed(() => {
